Move accounting exports to component with call for income events
